Moved Usermodel to App\Models, change namespace and calls to it.
